% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_utilities.R
\name{evaluate_models}
\alias{evaluate_models}
\title{Evaluate Models}
\usage{
evaluate_models(models, X_test_pca, Y_test)
}
\arguments{
\item{models}{A list of trained models.}

\item{X_test_pca}{Transformed testing data after PCA.}

\item{Y_test}{Actual target values for testing.}
}
\value{
A list containing evaluation metrics and predictions for each model.
}
\description{
Evaluates model performance using RMSE and R-squared metrics.
}
