% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_and_process_indicators.R,
%   R/merge_all_indicators.R
\name{fetch_and_process_indicators}
\alias{fetch_and_process_indicators}
\title{Fetch and Process Indicators}
\usage{
fetch_and_process_indicators(
  indicator_urls,
  num_cores = parallel::detectCores() - 1
)

fetch_and_process_indicators(
  indicator_urls,
  num_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{indicator_urls}{List of URLs for World Bank indicators.}

\item{num_cores}{Number of cores to use for parallel processing. Defaults to all available cores minus one.}
}
\value{
A list of tibbles with processed data for each indicator.

A list of tibbles with processed data for each indicator.
}
\description{
Fetch and process World Bank indicators using parallel processing.

Fetch and process World Bank indicators using parallel processing.
}
