% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_merge_indicators.R
\name{fetch_merge_indicators}
\alias{fetch_merge_indicators}
\title{Fetch and Merge Indicators}
\usage{
fetch_merge_indicators(
  indicator_urls,
  num_cores = parallel::detectCores() - 1,
  start_year = 2015,
  end_year = 2022
)
}
\arguments{
\item{indicator_urls}{A list of World Bank API URLs for indicators.}

\item{num_cores}{Number of cores for parallel processing. Default is all available cores minus one.}

\item{start_year}{The start year for filtering the data. Default is 2015.}

\item{end_year}{The end year for filtering the data. Default is 2022.}
}
\value{
A merged dataset containing all the processed indicators.
}
\description{
Fetches and processes multiple World Bank indicators and merges them into a single dataset.
}
