% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_paginated_data.R
\name{fetch_paginated_data}
\alias{fetch_paginated_data}
\title{Fetch Paginated Data}
\usage{
fetch_paginated_data(url)
}
\arguments{
\item{url}{A URL for a World Bank API endpoint.}
}
\value{
A tibble containing the combined paginated data.
}
\description{
Fetches paginated data from a given API URL.
}
