% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_all_indicators.R
\name{merge_all_indicators}
\alias{merge_all_indicators}
\title{Merge All Indicators}
\usage{
merge_all_indicators(indicator_data)
}
\arguments{
\item{indicator_data}{List of tibbles containing processed indicator data.}
}
\value{
A single merged tibble with all indicators.
}
\description{
Merge multiple indicator datasets into a single dataset.
}
