% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_utilities.R
\name{perform_pca}
\alias{perform_pca}
\title{Perform PCA}
\usage{
perform_pca(X_train, X_test, variance_threshold = 0.95)
}
\arguments{
\item{X_train}{Training data for PCA.}

\item{X_test}{Testing data for PCA.}

\item{variance_threshold}{The amount of variance to retain. Default is 0.95.}
}
\value{
A list containing the PCA model, transformed training data, and transformed testing data.
}
\description{
Applies Principal Component Analysis (PCA) to reduce dimensionality.
}
