% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_utilities.R
\name{plot_results}
\alias{plot_results}
\title{Plot Results}
\usage{
plot_results(Y_test, predictions_list, model_list)
}
\arguments{
\item{Y_test}{Actual target values for testing.}

\item{predictions_list}{A list of predictions from each model.}

\item{model_list}{A vector of model names to include in the plot.}
}
\value{
A ggplot object.
}
\description{
Generates a scatter plot comparing actual and predicted values for each model.
}
