% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_utilities.R
\name{split_data}
\alias{split_data}
\title{Split Data into Training and Testing Sets}
\usage{
split_data(X, Y, split_ratio = 0.8)
}
\arguments{
\item{X}{A matrix or data frame of predictors.}

\item{Y}{A vector of target values.}

\item{split_ratio}{The proportion of the data to use for training. Default is 0.8.}
}
\value{
A list containing training and testing datasets.
}
\description{
Splits predictors and target into training and testing datasets.
}
