% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_utilities.R
\name{train_models}
\alias{train_models}
\title{Train Models}
\usage{
train_models(X_train_pca, Y_train, model_list = c("lm", "rf", "svm", "gbm"))
}
\arguments{
\item{X_train_pca}{Transformed training data after PCA.}

\item{Y_train}{Target values for training.}

\item{model_list}{A vector of model names to train (e.g., "lm", "rf", "svm", "gbm").}
}
\value{
A list of trained models.
}
\description{
Trains specified machine learning models on the training data.
}
